Fix default paths being unsaveable
Almost all online sources say APFS (and HFS, etc) have 255 byte
limits on file names. And yet, practical testing reveals it to
be 251 on macOS 12 and 13 on an M1 mac.

Sigh.

This also adds a UTF8 truncator, since complex characters are
supported (e.g. é), which are multi-byte. Without it, we would
still be above our filepath limit