Free images from memory; fix History tab memory leak
The History tab was ever-increasing memory usage, and could
actually *slow down* image generation for large Histories.

--

Our History tab is rendered as a `v-if`, which means that its
component tree is dropped from memory when hidden, and is fully
rerendered when visible. That's generally a good thing, we don't
need all of that in memory while using other pages.

But it also means that the images themselves have to be refetched
each time, since Chromium does not apparently cache file:// images.
These new image entities get re-added to memory *every single time*
History is visited. That potentially balloons memory usage into
the tens of gigabytes as the app is used.

Downsides: this will make History more janky. Images will have to
reload every time it's visited.

That's OK - the solution to that is reducing the number of images
loaded at once, which will come later.